// src/app/interfaces/city.ts
export enum City {
  BlackMarket = 'Black Market',
  Bridgewatch = 'Bridgewatch',
  Caerleon = 'Caerleon',
  FortSterling = 'Fort Sterling',
  Lymhurst = 'Lymhurst',
  Martlock = 'Martlock',
  Thetford = 'Thetford',
  ArthursRest = 'Arthurs Rest',
  MorganasRest = 'Morganas Rest',
  MerlynsRest = 'Merlyns Rest'
}